+++ src/main/java/org/opensearch/searchrelevance/common/PluginConstants.java
@@ -49,6 +49,8 @@ public class PluginConstants {
     public static final String EVALUATION_RESULT_INDEX_MAPPING = "mappings/evaluation_result.json";
     public static final String JUDGMENT_CACHE_INDEX = ".plugins-search-relevance-judgment-cache";
     public static final String JUDGMENT_CACHE_INDEX_MAPPING = "mappings/judgment_cache.json";
+    public static final String EXPERIMENT_VARIANT_INDEX = "search-relevance-experiment-variant";
+    public static final String EXPERIMENT_VARIANT_INDEX_MAPPING = "mappings/experiment_variant.json";
 
     /**
      * UBI