+++ src/main/java/org/opensearch/searchrelevance/transport/experiment/PutExperimentTransportAction.java
@@ -9,7 +9,11 @@ package org.opensearch.searchrelevance.transport.experiment;
 
 import static org.opensearch.searchrelevance.common.MetricsConstants.METRICS_INDEX_AND_QUERIES_FIELD_NAME;
 import static org.opensearch.searchrelevance.common.MetricsConstants.METRICS_QUERY_TEXT_FIELD_NAME;
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_COMBINATION_TECHNIQUE;
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_NORMALIZATION_TECHNIQUE;
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_WEIGHTS_FOR_COMBINATION;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -29,15 +33,18 @@ import org.opensearch.common.inject.Inject;
 import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.rest.RestStatus;
 import org.opensearch.searchrelevance.dao.ExperimentDao;
+import org.opensearch.searchrelevance.dao.ExperimentVariantDao;
 import org.opensearch.searchrelevance.dao.QuerySetDao;
 import org.opensearch.searchrelevance.dao.SearchConfigurationDao;
 import org.opensearch.searchrelevance.exception.SearchRelevanceException;
 import org.opensearch.searchrelevance.experiment.ExperimentOptionsFactory;
 import org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch;
+import org.opensearch.searchrelevance.experiment.ExperimentVariantHybridSearchDTO;
 import org.opensearch.searchrelevance.metrics.MetricsHelper;
 import org.opensearch.searchrelevance.model.AsyncStatus;
 import org.opensearch.searchrelevance.model.Experiment;
 import org.opensearch.searchrelevance.model.ExperimentType;
+import org.opensearch.searchrelevance.model.ExperimentVariant;
 import org.opensearch.searchrelevance.utils.TimeUtils;
 import org.opensearch.tasks.Task;
 import org.opensearch.transport.TransportService;
@@ -49,6 +56,7 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
 
     private final ClusterService clusterService;
     private final ExperimentDao experimentDao;
+    private final ExperimentVariantDao experimentVariantDao;
     private final QuerySetDao querySetDao;
     private final SearchConfigurationDao searchConfigurationDao;
     private final MetricsHelper metricsHelper;
@@ -61,6 +69,7 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
         TransportService transportService,
         ActionFilters actionFilters,
         ExperimentDao experimentDao,
+        ExperimentVariantDao experimentVariantDao,
         QuerySetDao querySetDao,
         SearchConfigurationDao searchConfigurationDao,
         MetricsHelper metricsHelper
@@ -68,6 +77,7 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
         super(PutExperimentAction.NAME, transportService, actionFilters, PutExperimentRequest::new);
         this.clusterService = clusterService;
         this.experimentDao = experimentDao;
+        this.experimentVariantDao = experimentVariantDao;
         this.querySetDao = querySetDao;
         this.searchConfigurationDao = searchConfigurationDao;
         this.metricsHelper = metricsHelper;
@@ -84,9 +94,10 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
         String timestamp = TimeUtils.getTimestamp();
         Map<String, Object> results = new HashMap<>();
 
-        // step 1: Create Index if not exists
+        // step 1: Create Indexes if not exist
         StepListener<Void> createIndexStep = new StepListener<>();
         experimentDao.createIndexIfAbsent(createIndexStep);
+        experimentVariantDao.createIndexIfAbsent(createIndexStep);
 
         // step 2: Get QuerySet
         StepListener<Map<String, Object>> getQuerySetStep = new StepListener<>();
@@ -194,7 +205,7 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
                         error -> handleFailure(error, hasFailure, experimentId, request)
                     )
                 );
-            } else if (request.getType() == ExperimentType.HYBRID_SEARCH) {
+            } else if (request.getType() == ExperimentType.HYBRID_OPTIMIZER) {
                 Map<String, Object> defaultParametersForHybridSearch = ExperimentOptionsFactory
                     .createDefaultExperimentParametersForHybridSearch();
                 ExperimentOptionsForHybridSearch experimentOptionForHybridSearch =
@@ -202,6 +213,34 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
                         ExperimentOptionsFactory.HYBRID_SEARCH_EXPERIMENT_OPTIONS,
                         defaultParametersForHybridSearch
                     );
+                List<ExperimentVariantHybridSearchDTO> experimentVariantDTOs = experimentOptionForHybridSearch.getParameterCombinations(
+                    true
+                );
+                List<ExperimentVariant> experimentVariants = new ArrayList<>();
+                for (ExperimentVariantHybridSearchDTO experimentVariantDTO : experimentVariantDTOs) {
+                    Map<String, Object> parameters = new HashMap<>(
+                        Map.of(
+                            EXPERIMENT_OPTION_NORMALIZATION_TECHNIQUE,
+                            experimentVariantDTO.getNormalizationTechnique(),
+                            EXPERIMENT_OPTION_COMBINATION_TECHNIQUE,
+                            experimentVariantDTO.getCombinationTechnique(),
+                            EXPERIMENT_OPTION_WEIGHTS_FOR_COMBINATION,
+                            experimentVariantDTO.getQueryWeightsForCombination()
+                        )
+                    );
+                    String experimentVariantId = UUID.randomUUID().toString();
+                    ExperimentVariant experimentVariant = new ExperimentVariant(
+                        experimentVariantId,
+                        TimeUtils.getTimestamp(),
+                        ExperimentType.HYBRID_OPTIMIZER,
+                        AsyncStatus.PROCESSING,
+                        experimentId,
+                        parameters,
+                        Map.of()
+                    );
+                    experimentVariants.add(experimentVariant);
+                    experimentVariantDao.putExperimentVariant(experimentVariant, ActionListener.wrap(response -> {}, e -> {}));
+                }
                 metricsHelper.processEvaluationMetrics(
                     queryText,
                     indexAndQueries,
@@ -220,7 +259,7 @@ public class PutExperimentTransportAction extends HandledTransportAction<PutExpe
                             judgmentList
                         );
                     }, error -> handleFailure(error, hasFailure, experimentId, request)),
-                    experimentOptionForHybridSearch
+                    experimentVariants
                 );
             } else if (request.getType() == ExperimentType.POINTWISE_EVALUATION) {
                 metricsHelper.processEvaluationMetrics(