+++ src/test/java/org/opensearch/searchrelevance/experiment/QuerySourceUtilTests.java
@@ -7,18 +7,23 @@
  */
 package org.opensearch.searchrelevance.experiment;
 
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_COMBINATION_TECHNIQUE;
+import static org.opensearch.searchrelevance.experiment.ExperimentOptionsForHybridSearch.EXPERIMENT_OPTION_NORMALIZATION_TECHNIQUE;
+
 import java.util.List;
 import java.util.Map;
 
+import org.opensearch.searchrelevance.model.ExperimentVariant;
 import org.opensearch.test.OpenSearchTestCase;
 
 public class QuerySourceUtilTests extends OpenSearchTestCase {
 
     public void testCreateDefinitionOfTemporarySearchPipeline_ValidInput_ReturnsCorrectStructure() {
         // Given
-        SubExperimentHybridSearchDao experimentHybridSearchDao = SubExperimentHybridSearchDao.builder()
-            .normalizationTechnique("min_max")
-            .combinationTechnique("arithmetic_mean")
+        ExperimentVariant experimentHybridSearchDao = ExperimentVariant.builder()
+            .parameters(
+                Map.of(EXPERIMENT_OPTION_NORMALIZATION_TECHNIQUE, "min_max", EXPERIMENT_OPTION_COMBINATION_TECHNIQUE, "arithmetic_mean")
+            )
             .build();
 
         // When